Refactor scheduler implementation to use ScheduledRunnable and improve blocking logic
